Copy remote_response headers into proxy_response
Copies headers all headers from the remote_response to the response
that is ultimately output by the proxy. Some of these are required
such as 'Content-Type' and 'Content-Length' so that clients will be
able to correctly read the response.

Currently, firefox has some issues receiving the response from the
reverse_proxy and will try to download a file instead of rendering
text even though the output of the reverse_proxy matches the
regular output when requested by curl. Also, the reverse_proxy
fails to transfer files (bunny.jpg) and for some reason stops after
sending 4 bytes.