Remove HandleRequest test case from unit tests
This is not suitable for a unit test, as it would
involve actually making a request to ucla.edu
(since we don't have mocks to use). Unit tests
are meant to be exercises of self-contained code that don't
require use of external resources. This would be better
suited for an integration test.